render msg.asc for multipart/encrypted mails
this is the first step to support reading PGP encrypted mails using
mailvelope browser extension. Works quite nice, but the result is a
multipart message rendered in plain-text by mailvelope. It would require
multipart parsing to be compatible with other implementations of PGP.
Anyway, this is out of the scope of this plugin and would need to be
done in mailvelope.