make editor an event hub and listen for editorMode change events
this can be used to register for different events trigged by the mail
model or any other object from the mail compose dialog.

The first use-case is to listen for editorMode change events and force
an update of the tokenfields, in case. So the key lookup can be
triggered, again.